requirements.txt
streamlit>=1.43.1
openai>=1.0.0
sentence-transformers>=2.2.2
pymongo>=4.6.1
python-dotenv>=1.0.1
numpy>=1.24.0
torch>=2.0.0 